+++ cli/runtime/commands/run/module.go
@@ -6,20 +6,20 @@ import (
 	"os"
 
 	"github.com/wago-org/wago"
+	"github.com/wago-org/wago/artifactcache"
 	"github.com/wago-org/wago/cli/internal/ui"
 	"github.com/wago-org/wago/cli/internal/wasmcall"
-	"github.com/wago-org/wago/cli/runtime/internal/artifactcache"
 )
 
-func mustLoadModule(file string, config *wago.RuntimeConfig, runtime *wago.Runtime, cache artifactcache.Cache, allowNativeArtifact bool) *wago.Module {
-	module, err := loadModule(file, config, runtime, cache, allowNativeArtifact)
+func mustLoadModule(file string, runtime *wago.Runtime, cache artifactcache.Cache, allowNativeArtifact bool) *wago.Module {
+	module, err := loadModule(file, runtime, cache, allowNativeArtifact)
 	if err != nil {
 		ui.Fatal("%v", err)
 	}
 	return module
 }
 
-func loadModule(file string, config *wago.RuntimeConfig, runtime *wago.Runtime, cache artifactcache.Cache, allowNativeArtifact bool) (*wago.Module, error) {
+func loadModule(file string, runtime *wago.Runtime, cache artifactcache.Cache, allowNativeArtifact bool) (*wago.Module, error) {
 	source, err := os.ReadFile(file)
 	if err != nil {
 		return nil, err
@@ -39,7 +39,7 @@ func loadModule(file string, config *wago.RuntimeConfig, runtime *wago.Runtime, 
 		}
 		return module, nil
 	}
-	module, err := cache.LoadOrCompile(source, config, runtime)
+	module, err := cache.LoadOrCompile(source, runtime)
 	if err != nil {
 		return nil, err
 	}